+++ lib/DTO/GrantOrTokenRequestArgs.php
@@ -1,10 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace OpenPayments\DTO;
 
 class GrantOrTokenRequestArgs extends UnauthenticatedResourceRequestArgs
 {
-    public string $accessToken;
+    public readonly string $accessToken;
 
     public function __construct(string $url, string $accessToken)
     {