+++ lib/DTO/PendingGrantAccessToken.php
@@ -1,10 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace OpenPayments\DTO;
 
 class PendingGrantAccessToken
 {
-    public string $value;
+    public readonly string $value;
 
     /**
      * @param  string  $value  The value of the continuation access token.